Setup unit tests infrastructure and add header tests
This commit installs all the necessary libraries such as react testing library and jest-dom to allow unit tests to be
created for components. The configuration files such as SetupTests and jest.config sort out the infrastructure needed to get the tests running.
Finally tests for the header has been added.