Feature: Project details page (#1)
* WIP: First pass for Project details view

* Add packages for important functionality

A couple of packages are added within this commit. packages include axios for API calls, classnames for conditional classes for components,
http-status-codes which set common HTTP status codes as constants, reack-hook-form for future form creation and swr to ease fetching for up-to-date data

* Add mock factory for project and add new type

This commit adds a mock factory for a project object for ease of testing. The project type is also updated to include further information and a generic field type has been added for future form components

* Create components for project details page

The main of components has been created for the project details page such as tables to display information and a view wrapper component to allow all children components to have access to project data.

* Update project details page, add mock api calls